cloud_function_src/vectorization-delete/requirements.txt
functions-framework==3.*
pg8000
asyncio
asyncpg
get_metadata
cloud-sql-python-connector
google-cloud-secret-manager